queue_url_holder
